+++ inst/shinyexample/dev/launch_run_dev.R
@@ -1,0 +1,7 @@
+# You can choose your run_dev if choose = TRUE
+my_app <- golem::golemLaunch$new()
+
+# More actions
+my_app$open_app()
+
+# my_app$stop()